Add async DmxInput capability.
DmxInput is now running in the background,
using a DMA channel and the DMA_IRQ0 interrupt.

This approach gives the following advantages:
- Supports both blocking and non-blocking use
- CPU knows when a packet was received
- Supports any number of channels (not just powers of 2)
- Requires no memory alignment of the buffer

An alternative method using the DMA ring feature was also
considered. This method would have meant almost no cpu load
with just one interrupt every 2**32 received channels.
However, this would require the number of channels
to be a power of two, the buffer would need to be naturally aligned
(256 channels would need a 256-byte aligned buffer), and there would be
no way to tell if dmx signals have stopped coming in.